@@ -1,16 +1,23 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import './index.css';
 import App from './App';
 import './i18n';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
+import 'bootstrap/dist/css/bootstrap.min.css';
+
+import './i18n';
+
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement,
+);
+root.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root'),
 );
 
 // If you want to start measuring performance in your app, pass a function
