@@ -1,6 +1,12 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
-import i18n from '../src/i18n';
+import '../src/i18n';
 
 export const parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
+  controls: {
+    matchers: {
+      color: /(background|color)$/i,
+      date: /Date$/,
+    },
+  },
 };
