@@ -1,7 +1,7 @@
 import React from 'react';
 import './button.css';
 
-export interface ButtonProps {
+interface ButtonProps {
   /**
    * Is this the principal call to action on the page?
    */
@@ -27,22 +27,18 @@ export interface ButtonProps {
 /**
  * Primary UI component for user interaction
  */
-export const Button: React.FC<ButtonProps> = ({
+export const Button = ({
   primary = false,
   size = 'medium',
   backgroundColor,
   label,
   ...props
-}) => {
-  const mode = primary
-    ? 'storybook-button--primary'
-    : 'storybook-button--secondary';
+}: ButtonProps) => {
+  const mode = primary ? 'storybook-button--primary' : 'storybook-button--secondary';
   return (
     <button
       type="button"
-      className={['storybook-button', `storybook-button--${size}`, mode].join(
-        ' ',
-      )}
+      className={['storybook-button', `storybook-button--${size}`, mode].join(' ')}
       style={{ backgroundColor }}
       {...props}
     >
