@@ -1,6 +1,6 @@
 export interface Panda {
-  key?: string;
+  key: string;
   name: string;
   image: string;
-  interests: string[];
+  interests: Array<string>;
 }
